test: test case 추가 및 test 분리 NetworkManager의 test와 JSONDecoder의 test로 파일 분리 NetworkManager에서 PUSH, PATCH, DELETE 메소드에 대한 테스트 코드 추가
